Changing the way messages are handled.
Previously, we did not TranslateMessage. This did not work well with
the newly added textbox on the main window. What happened was that
because the keyboard input was not put into the message queue, the
edit control never received anything. It wasn't until TranslateMessage
was added that the edit control was usable.

However, after TranslateMessage was added, any keys typed will result
in multiple characters sent to the screen. This change now returns
from all WM_KEYDOWN and WM_SYSKEYDOWN (also their UP variants) and
prevents further handling of the message. This allows the handling of
valid keyboard characters in WM_CHAR, thus it does not result in multiple
characters being sent to the screen.