TranslateMessage fix earlier didn't work.
I had assumed that the code was not doing TranslateMessage when in
fact it does inside of win_key_down. This change reverts what was done
previously and now checks to see if the edit control has focus. If it
does, then execute TranslateMessage so that the message gets through
to the control.