Added hotkey support, proper painting, and did some cleaning.
The hotkey support now hooks into the Alt+F key. There was additional
subclassing of the edit control to get it to handle Alt+F and Escape
when it is focused. This means that the search box can now show/hide
at will :).

There is now proper painting support for the search control box. When
it is visible, its rectangle is clipped out of the update function. This
is necessary so that the textbox remains visible when the terminal is
in focus or being highlighted.

Some cleaning and refactoring of code.