Adding a different highlight for search.
Added configuration options for the search fg/bg color. These colors
are now used to highlight the text that matches the search query.
As a result, there's now an ATTR_SEARCH attribute for the characters.

Also fixed a small bug when hiding the search control and the terminal
does not scroll to the bottom.