Adding (shift)enter to search previous and next.
Now pressing the enter key in the search control box will bring you
to the previous result. Since we are going backward from the bottom
of the terminal up, each enter key will search upward into the
scrollback.

This additional behavior requires that the current search result that
is being centered in the screen have a different background to
differentiate itself from fellow results. Now, we have a third option
in the configuration to set the background color for the current
search result that is being centered on.